added routing to calculators page and moved cards
