+++ src/pages/Calculators.js
@@ -1,38 +1,36 @@
-import Card from 'components/Card'
+import AllCalcs from 'calculators/AllCalcs'
+import SellBuyBack from 'calculators/SellBuyBack'
 import Hero from 'components/Hero'
 import PageWrapper from 'layouts/PageWrapper'
 import React from 'react'
+import {
+    BrowserRouter as Router,
+    Route,
+    Switch,
+    useRouteMatch
+} from 'react-router-dom'
 
 const Calculators = () => {
 
-    const listOfCalcs = [
-        {
-            title: 'Sell The Top and Buy Back',
-            description: 'Think you can time the top and buy back in? Find out how much the price would have to drop for you to end up with the same amount of bitcoin and how well you would fair under different scenarios.'
-        },
-        {
-            title: 'Retire on Bitcoin',
-            description: 'Figure out how much bitcoin you would need to retire'
-        }
-    ]
-
-    listOfCalcs.map(calc => console.log(calc.title))
-
-
+    let match = useRouteMatch()
 
     return (
-        <PageWrapper>
-            <Hero
-                title='Calculators'
-                subtitle='list of calculators'
-            />
-            {listOfCalcs.map(calc =>
-                <Card
-                    title={calc.title}
-                    description={calc.description}
+        <Router>
+            <PageWrapper>
+                <Hero
+                    title='Calculators'
+                    subtitle='list of calculators'
                 />
-            )}
-        </PageWrapper>
+                <Switch>
+                    <Route exact path={`${match.path}/`}>
+                        <AllCalcs />
+                    </Route>
+                    <Route path={`${match.path}/hello`}>
+                        <SellBuyBack />
+                    </Route>
+                </Switch>
+            </PageWrapper>
+        </Router>
     )
 }
